Use a mutex to prevent DLL to be loaded twice
Previously the exe's name would have been checked, this could lead to problems if the user has the exe renamed to something else. A mutex should prevent the problem with naming, Win32 docs recommend to create a locked file in the user's profile directory, but for now this should be enough.